moved 'home' to it's own file, cleaned up code
